6782 bytes

+++ firmware/src/firmware_packet_process.c
@@ -93,8 +93,8 @@ void treat_write( struct connection * c, const struct packet * p )
   else {
     mask m = new_mask( c->caps.nb_pins );
     read_mask( p->data, 0, m, c->caps.nb_pins );
-    int pin_index = 0;
-    unsigned int offset = c->caps.nb_pins;
+    uint8_t pin_index = 0;
+    uint16_t offset = c->caps.nb_pins;
     do{
       pin_index = mask_next_pin_used( m, pin_index, c->caps.nb_pins );
       if ( pin_index == -1 ) break;
